Model implementation for consultation response forms.
Some consultations can offer a form for download as a means of
responding. This is the model implementation to achieve that.

I've kept with the pattern of including a bridge model between the
edition and the actual file, so that forms can be removed from later
editions without affecting earlier ones, and the title of the form
can also be modified on a per-edition basis.