Fix bug in `Speech#populate_organisations_based_on_role_appointment`.
The data migration in [1] failed to update the organisations on each speech
to match its new role appointment, because of a bug in
`Speech#populate_organisations_based_on_role_appointment` which is
invoked using a `before_save` callback on `Speech`.

This commit introduces a test in `SpeechTest` to demonstrate the bug and
fixes it by changing the implementation of the callback to make use of
standard ActiveRecord functionality to create & destroy join table
records for a `has_many :through` relationship (i.e. by assigning
`Speech#organisations` in this case).

I found I also had to set `Speech#edition_organisations` to an empty
Array in order to avoid inserting duplicate join table records in some
tests. This was because
`Edition::Organisations::Trait#process_associations_before_save`
manipulates `Edition#edition_organisations` which doesn't play well with
also manipulating `Edition#organisations` before the edition is saved.

Lastly, I had to update tests that we attempting to assign organisations
to speeches built by factories so that the speech was actually assigned
a role appointment with a role in the organisation. This actually seems
more realistic.

[1] 0714881bf84ec73b953da0543cfcc47403478a66