List URLs on which a publication appears.
This is so we can purge the cache for these URLs when a document is
published at a scheduled time and must appear across the site
immediately.

Note that I haven't yet tackled the problem of where we're going to
obtain the host, but this is ok for the moment, because we're still
using a fake cache buster.