List URLs on which news articles and speeches appear.
Note that this behaviour is almost the same for news articles as it is
for speeches, so I've implemented most of it on the common parent class,
`Announcement`.

Speeches are a slightly special case, because they are potentially
associated to a single minister page via their single role appointment.

Although I don't think the annoucements page is available as an atom
feed at the moment, the route exists and I thought it might future-proof
things a bit if we included it in the list. It also makes things more
consistent with the `Publicationesque` behaviour.