List URLs on which policies appear.
As with previous commits this is so we can do some cache-busting
for scheduled publishing.

There is definitely some duplication creeping in, but I plan to address
that in a later commit.

I'm not sure that an atom feed for policies is currently made available,
but since the route exists, I thought it should be included.