Task to show editions with missing attachments
Originally we were going to run this to correct edition attachment data
for latests editions on production, but we discovered there were only 12
affected documents, so we decided to fix them manually. This is checked
in and left here for reference. Running it is non-desctructive at
present as it just creates new records in memory and prints some
diagnostic output.

See the discussion on the ticket here:
https://www.pivotaltracker.com/story/show/39402077