Reduce cache expiry for 404 on publication scheduled for publication.
When a publication is scheduled for publication, the cache control
max-age for index pages that will include the publication is
reduced below the default as the scheduled publication time
approaches in order that there will be cache misses for the index
pages at the scheduled publication time and the up-to-date content will
be seen immediately. This is all working ok.

However, if someone hits the planned public URL of the publication
within 30 minutes (the default expiry time) of the scheduled publication
time and (correctly) gets a 404 response, this 404 response will be
cached by Akamai and/or Varnish.

In order to resolve this, we need to set the cache control max-age
header on these 404s using the same logic as for the index pages
described above. This should mean that although the 404 would still be
cached, the cache will expire in time to generate a cache miss at the
scheduled publication time and the up-to-date content should be
displayed as intended.

For the moment, I have only applied this change to publications, but it
needs to be applied to all document types that can be scheduled for
publication.