Experimental - authenticate public facing requests on admin
NOT WORKING

The aim of this is to allow us to remove basic auth from the admin host.
If we do this, we'd want to require SSO authentication on the
admin host. However, although it's easy to add the before_filter to the
public facing controller, I had trouble keeping the tests working.

The problem was that when running testsauthentication failed in the
authenticate_user! call it didn't halt the request (in the tests).