Spike implementing translation at attribute level.
This general approach should technically work for any model, not just
editions.

Implementation
-----

Translations are stored in `TranslatedAttribute` instances, which can
be associated with any object as a polymorphic 'owner'.

In Edition, attributes with translations are declared:

    translate_attributes :title, :body, :summary

This declares methods which will look up the contents of those
attributes from an instance cache, according to the value of the
instance locale. If attributes are looked up with no locale, or with
the default locale, the attribute cache is not populated or loaded,
so there should be no impact to performance. If a locale is set, all
translated attributes for that instance are loaded in a single query.

Frontend display
--------

The instance locale is set when the model is loaded in the controller

    find_document_or_edition.for_locale(locale)

The `for_locale` method sets the `@current_locale` of that instance,
and returns `self`. The `locale` parameter is a method available on
`ApplicationController`, which can be trivially set at the moment by
adding a query parameter `locale=es-ES` to any URL.

With this approach, no views need to be modified; because each
instance knows "within itself" what locale to use, the translated
content will appear in place.

Workflow
-------

When a new edition is created, all translated attributes are also
duplicated, so that they can be edited without affecting published
versions.

This means that it's (at least currently) possible to publish with
only a few attributes translated.

We'd need to think more about how to handle that.

Admin
-----

In the admin, the inputs for the translated attributes are exposed
as a nested form, and the Edition accepts attributes for translated
attributes:

    accepts_nested_attributes_for :translated_attributes

I've used some javascript to bring the attribute forms together, with
a little link to switch between different translations on a per-input
basis, although I'd almost certainly do something a bit more
comprehensive in a real implementation (i.e. a single control to
switch all inputs to a particular translation locale).

It's not yet possible to add a *new* translation locale with this
javascript in place, but it shouldn't be hard to extend this approach;
just a bit of a grind.