Spike-out some code to allow non-English editions
Because we can no longer assume that an edition is available in
English, we have to work a bit harder to make sure we load editions
in a locale for which we have content for. Here, I've introduced the
concept of a primary locale on Edition. This will be :en if an
English translation is available, or the first available translation
otherwise. The primary locale is used to set up a LocalisedModel of
edition and to make sure we link to the correct version of the
edition.

Notice too that available_in_multiple_languages? has been overriden
for WorldLocationNewsArticle. This is because the original definition
in TranslatableModel makes the assumption that an English version
will always be present.