Enable Foreign Secs historical accounts in admin
This allows content to start adding historical accounts for Foreign
Secretaries so that we can make it dynamic rather than static. The
data migration is the main thing that allows historical accounts to be
created for people in the role of Foreign Secretary. The routing
changes are required for two reasons:

1. So the Foreign Secretary public role page can generate the link
to the historical role page

2. So the historical roles page in admin can link to "View" a given
historical account.

Note that the "View" link in admin will not work in all cases because
there will be some slug mismatches between the current static pages
and the new dynamic ones. For example, Robert Gascoyne-Cecil 3rd
Marquess of Salisbury is already a person with the slug
"robert-gascoyne-cecil", but his current static historical account
page uses the slug "robert-cecil". This will have to be delt with in
the next story that switches to the dynamic version of the public
pages.