World org main offices need to be treated differently
They can go in the list, but when we ask for the offices in the list remove it from that list.  Also say it's on the home page if asked and it's not in the list ('cos it will be on the home page).  Tweak the admin UI to not allow ordering the main office, but do show it as being first.